mods from bertrand #1997

+++ radio/src/bootloader/boot.cpp
@@ -193,11 +193,11 @@ void interrupt10ms(void)
 {
   Tenms |= 1;			// 10 mS has passed
 
-  uint8_t enuk = KEY_MENU;
-  uint8_t in = ~readKeys();
+  uint16_t enuk = KEY_MENU;
+  uint16_t in = ~readKeys();
 
   for (int i = 1; i < 7; i++) {
-    uint8_t value = in & (1 << i);
+    uint16_t value = in & (1 << i);
     keys[enuk].input(value, (EnumKeys) enuk);
     ++enuk;
   }
@@ -460,8 +460,8 @@ void writeEepromBlock()
 
 int main()
 {
-  uint8_t index = 0;
-  uint8_t maxhsize = DISPLAY_CHAR_WIDTH;
+  uint32_t index = 0;
+  uint32_t maxhsize = DISPLAY_CHAR_WIDTH;
   FRESULT fr;
   uint32_t state = ST_START;
   uint32_t nameCount = 0;
